Add files via upload
Simplilearn
Project: Developing a Backend Admin for Learner’s Academy.
DESCRIPTION
Project objective: 
As a Full Stack Developer, design and develop a backend administrative portal for the Learner’s Academy. Use the GitHub repository to manage the project artifacts. 
 
Background of the problem statement:
Learner’s Academy is a school that has an online management system. The system keeps track of its classes, subjects, students, and teachers. It has a back-office application with a single administrator login.
 
The administrator can:
● Set up a master list of all the subjects for all the classes
● Set up a master list of all the teachers
● Set up a master list of all the classes
● Assign classes for subjects from the master list
● Assign teachers to a class for a subject (A teacher can be assigned to different classes for different subjects)
● Get a master list of students (Each student must be assigned to a single class)
     
There will be an option to view a Class Report which will show all the information about the class, such as the list of students, subjects, and teachers
     
The goal of the company is to deliver a high-end quality product as early as possible. 
The flow and features of the application: 
● Plan more than two sprints to complete the application
● Document the flow of the application and prepare a flow chart 
● List the core concepts and algorithms being used to complete this application
● Implement the appropriate concepts, such as exceptions, collections, and sorting techniques for source code optimization and increased performance